Move attributes equality fn to utils
